And new sSHT interface, pixelizations, λiterator, and simpler evaluation of sYlm (#15)
The primary goal of this commit is introduction of the new sSHT interface, which looks somewhat like the AbstractFFTs interface, where the transformation objects can simply be multiplied by the vector to transform as in f = 𝒯 * f̃, or left-divided to reverse the transformation as in f̃ = 𝒯 \ f. This implements something like the Reinecke-Seljebotn algorithm, my modified version of the EKKM algorithm, and my direct algorithm.